+++ SS31.Client/Input/Event.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Input;
 
-namespace SS31.Common
+namespace SS31.Client.Input
 {
 	// Delegate for KeyPressed, KeyReleased, KeyHeld
 	public delegate void KeyEvent(Keys key, KeyboardState current, KeyboardState last);
